aws-lambda-classifier/requirements.txt

sentence-transformers==2.2.2
s3fs==2023.10.0
